Move PushButton to common lib
